complete set share operate method

+++ manage/Controller/ShareToolController.php
@@ -90,6 +90,8 @@ class ShareToolController extends AppController {
                     'scope_id' => $share_id,
                     'scope_type' => SHARE_OPERATE_SCOPE_TYPE);
                 $this->ShareOperateSetting->save($saveData);
+                Cache::write(SHARE_ORDER_TAG_OPERATE_CACHE_KEY . '_' . $share_id, '');
+                Cache::write(SHARE_ORDER_TAG_OPERATE_CACHE_KEY . '_' . $share_id . '_' . $tag_id, '');
             }
         }
     }
@@ -112,6 +114,7 @@ class ShareToolController extends AppController {
                     'scope_id' => $share_id,
                     'scope_type' => SHARE_OPERATE_SCOPE_TYPE);
                 $this->ShareOperateSetting->save($saveData);
+                Cache::write(SHARE_ORDER_OPERATE_CACHE_KEY . '_' . $share_id, '');
             }
         }
     }
@@ -120,10 +123,20 @@ class ShareToolController extends AppController {
     /**
      * @param $id
      * @param $share_id
+     * @param $data_id
      * 删除分享权限
      */
-    public function admin_delete_share_operate_setting($id, $share_id) {
-        $this->ShareOperateSetting->delete($id);
+    public function admin_delete_share_operate_setting($id, $share_id, $data_id) {
+        $data = $this->ShareOperateSetting->find('first', array('conditions' => array('id' => $id)));
+        if (!empty($data)) {
+            $this->ShareOperateSetting->delete($id);
+            if ($data['ShareOperateSetting']['data_type'] == SHARE_ORDER_OPERATE_TYPE) {
+                Cache::write(SHARE_ORDER_OPERATE_CACHE_KEY . '_' . $share_id, '');
+            } else {
+                Cache::write(SHARE_ORDER_TAG_OPERATE_CACHE_KEY . '_' . $share_id, '');
+                Cache::write(SHARE_ORDER_TAG_OPERATE_CACHE_KEY . '_' . $share_id . '_' . $data_id, '');
+            }
+        }
         $this->redirect(array('action' => 'admin_share_operate_set_view', '?' => array('share_id' => $share_id)));
     }
